refactor: 대시보드 도메인 구조 정리 및 DTO 위치 재배치
